requirements-locked.txt
# Development lock (minimal, illustrative). Prefer uv/pip-tools for real pins.

basicsr==1.4.2
facexlib==0.3.0
opencv-python==4.12.0.88
scikit-image==0.25.2
scipy==1.16.2
numpy==2.3.3
tqdm==4.67.1
huggingface_hub==0.34.5
pytest-cov==5.0.0

# API
fastapi==0.116.2
uvicorn[standard]==0.35.0
python-multipart==0.0.20
slowapi==0.1.9
piexif==1.1.3

# Metrics (optional)
lpips==0.1.4
